Fix integer overflow in TFLite tensor dimension calculations.
Reverts changelist 930010564

FUTURE_COPYBARA_INTEGRATE_REVIEW=https://github.com/openxla/xla/pull/43742 from ROCm:feature/pr1-collective-kernel-strategy-annotator 84543429efde2c72b77eb1797ff8741e02d747af
PiperOrigin-RevId: 930195515

+++ tensorflow/lite/delegates/gpu/api.cc
@@ -105,7 +105,7 @@ bool IsObjectInitialized(const TensorObject& obj) {
   return GetType(obj) != ObjectType::UNKNOWN;
 }
 
-uint32_t NumElements(const TensorObjectDef& def) {
+int64_t NumElements(const TensorObjectDef& def) {
   const auto& d = def.dimensions;
   switch (def.object_def.data_layout) {
     case DataLayout::BHWC:
@@ -113,7 +113,7 @@ uint32_t NumElements(const TensorObjectDef& def) {
     case DataLayout::HWDC4:
     case DataLayout::HDWC4:
     case DataLayout::DHWC4:
-      return d.b * d.h * d.w * AlignByN(d.c, 4);
+      return static_cast<int64_t>(d.b) * d.h * d.w * AlignByN(d.c, 4);
     case DataLayout::UNKNOWN:
       return 0;
   }
